todo en plus partie.java

+++ JeuTuilesGenerique/Modele/Partie.java
@@ -93,7 +93,7 @@ public class Partie implements Serializable {
         // TODO lancement d'un thread timer (en background) pour ne pas bloquer l'interface
         gui.activerBoutonsTuileAJouer();
         
-        int[] meilleureTuile = recursiveIA(plateau.largeur/2, plateau.hauteur/2, new ArrayList<Tuile>()); 
+        int[] meilleureTuile = recursiveIA((int) plateau.largeur/2, (int) plateau.hauteur/2, new ArrayList<Tuile>()); 
         // Le tableau meilleureTuile resterait null dans le cas où tuileAJouer n'est plaçable nulle part.
         if (meilleureTuile != null) {
             for (int i = 0; i < meilleureTuile[2]; i++) {
@@ -111,9 +111,10 @@ public class Partie implements Serializable {
          // Pour ne pas tester 2 fois la même tuile et donc tomber dans une boucle infinie.
         if (tuilesTestees.contains(plateau.plateau[x][y])) return null;
         tuilesTestees.add(plateau.plateau[x][y]) ;
+        // TODO optimiser IA en retournant directement null si la tuile en x,y est vide et ses voisines aussi.
         int[] meilleureTuile = null;
-        // Teste si on peut passer la tuileAJouer aux coordonnées passées en argument dans les 4 sens possibles.
-        for (int i = 0; i < 4; i++)  {
+        // Teste si la tuileAJouer peut être posée aux coordonnées passées en argument dans les 4 sens possibles.
+        for (int i = 0; i < 4; i++) {
             if (check(x, y)) {
                 int pointsRapportes = nbPoints(x, y);
                 if (meilleureTuile == null || pointsRapportes > meilleureTuile[3]) meilleureTuile = new int[]{x, y, i, pointsRapportes};
